chore: move Wallet to its own dir
